Refactor lambda - move messageformatter - simplify logic - update tests (less internal implementation details required)
