+++ src/LogLambda/test/alarm.test.ts
@@ -1,6 +1,18 @@
 import fs from 'fs';
 import path from 'path';
-import { getEventType, parseMessageFromEvent, slackParamsFromMessage, createMessage, messageShouldTriggerAlert } from '../index';
+import axios from 'axios';
+import MockAdapter from 'axios-mock-adapter';
+import { getEventType, parseMessageFromEvent, messageShouldTriggerAlert, sendMessageToSlack, slackMessageFromSNSMessage } from '../index';
+
+let axiosMock: MockAdapter;
+beforeAll(() => {
+  process.env.SLACK_WEBHOOK_URL = 'http://nothing.test';
+  axiosMock = new MockAdapter(axios);
+});
+
+beforeEach(() => {
+  axiosMock.reset();
+});
 
 describe('Test message types', () => {
   test('ECS Task state change', async () => {
@@ -15,7 +27,7 @@ describe('Test message types', () => {
     const sampleAlarmEventJson = await getStringFromFilePath(path.join('samples', 'unknown-event.json'));
     const event = JSON.parse(sampleAlarmEventJson);
     const message = parseMessageFromEvent(event);
-    expect(() => slackParamsFromMessage(message)).toThrow();
+    expect(() => slackMessageFromSNSMessage(message)).toThrow();
   });
 });
 
@@ -52,23 +64,14 @@ describe('Test alarm state changes', () => {
   });
 
   test('Get values for alarm type message', async () => {
+    axiosMock.onPost().reply(200, {});
     const sampleAlarmEventJson = await getStringFromFilePath(path.join('samples', 'alarm.json'));
     const event = JSON.parse(sampleAlarmEventJson);
     const message = parseMessageFromEvent(event);
-    const values = slackParamsFromMessage(message);
-    expect(values.title).toBe('❗️ Alarm: Certificate about to expire');
-  });
-
-  test('Get slack message object', async () => {
-    let messageObject = {
-      title: '❗️ Alarm: Certificate about to expire',
-      message: 'Threshold Crossed: 1 out of the last 1 datapoints [227.0 (08/08/22 12:32:00)] was less than or equal to the threshold (360.0) (minimum 1 datapoint for OK -> ALARM transition).',
-      context: 'CloudWatch Alarm State Change',
-      url: 'https://eu-west-1.console.aws.amazon.com/cloudwatch/home?region=eu-west-1#alarmsV2:alarm/CloudWatch%20Alarm%20State%20Change',
-      url_text: 'Bekijk alarm',
-    };
-    const message = createMessage(messageObject);
-    expect(message.blocks).toHaveLength(4);
+    await sendMessageToSlack(slackMessageFromSNSMessage(message));
+    expect(axiosMock.history.post.length).toBe(1);
+    const data = JSON.parse(axiosMock.history.post[0].data);
+    expect(data?.blocks[0].text.text).toBe('❗️ Alarm: Certificate about to expire');
   });
 });
 
@@ -88,19 +91,23 @@ describe('ECS State changes', () => {
   });
 
   test('Get slack message object', async () => {
+    axiosMock.onPost().reply(200, {});
     const sampleEventJson = await getStringFromFilePath(path.join('samples', 'ecs-task-state-change.json'));
     const event = JSON.parse(sampleEventJson);
     const message = parseMessageFromEvent(event);
-    expect(slackParamsFromMessage(message).message).toBe('Containers involved: \\n - test-sleep (PENDING)');
+    await sendMessageToSlack(slackMessageFromSNSMessage(message));
+    const data = JSON.parse(axiosMock.history.post[0].data);
+    expect(data?.blocks[2].text.text).toBe('Containers involved: \n - test-sleep (PENDING)');
   });
 
   test('Get slack message object', async () => {
+    axiosMock.onPost().reply(200, {});
     const sampleEventJson = await getStringFromFilePath(path.join('samples', 'ecs-task-state-change.json'));
     const event = JSON.parse(sampleEventJson);
     const message = parseMessageFromEvent(event);
-    const params = slackParamsFromMessage(message);
-    const slackMessage = createMessage(params);
-    expect(slackMessage.blocks).toHaveLength(4);
+    await sendMessageToSlack(slackMessageFromSNSMessage(message));
+    const data = JSON.parse(axiosMock.history.post[0].data);
+    expect(data?.blocks).toHaveLength(4);
   });
 });
